+++ src/compas/datastructures/network/core/graph.py
@@ -5,9 +5,6 @@ from __future__ import division
 from random import sample
 from random import choice
 from ast import literal_eval
-from distutils.version import LooseVersion
-
-import compas
 
 from compas.datastructures.datastructure import Datastructure
 from compas.datastructures.attributes import NodeAttributeView
@@ -51,68 +48,9 @@ class Graph(Datastructure):
     >>>
     """
 
-    @property
-    def dataschema(self):
-        import schema
-        version = LooseVersion(compas.__version__)
-        meta = {
-            "compas": str,
-            "datatype": str,
-            "data": None
-        }
-        data = {
-            "attributes": dict,
-            "node_attributes": dict,
-            "edge_attributes": dict,
-            "node": dict,
-            "edge": dict,
-            "adjacency": dict,
-            "max_int_key": schema.And(int, lambda x: x >= -1)
-        }
-        if version < LooseVersion('0.16.5'):
-            return schema.Schema(data)
-        meta["data"] = data
-        return schema.Schema(meta)
-
-    @property
-    def jsonschema(self):
-        version = LooseVersion(compas.__version__)
-        schema = {
-            "$schema": "http://json-schema.org/draft-07/schema#",
-            "$id": "https://github.com/compas-dev/compas/schemas/graph.json",
-            "$compas": version.vstring.split('-')[0],
-        }
-        data = {
-            "type": "object",
-            "properties": {
-                "attributes":      {"type": "object"},
-                "node_attributes": {"type": "object"},
-                "edge_attributes": {"type": "object"},
-                "node":            {"type": "object"},
-                "edge":            {"type": "object"},
-                "adjacency":       {"type": "object"},
-                "max_int_key":     {"type": "number"}
-            },
-            "required": ["attributes", "node_attributes", "edge_attributes", "node", "edge", "adjacency", "max_int_key"]
-        }
-        if version < LooseVersion('0.16.5'):
-            schema.update(data)
-        else:
-            meta = {
-                "type": "object",
-                "properties": {
-                    "compas": {"type": "string"},
-                    "datatype": {"type": "string"},
-                    "data": data
-                },
-                "required": ["compas", "datatype", "data"]
-            }
-            schema.update(meta)
-        return schema
-
     def __init__(self):
         super(Graph, self).__init__()
-        self._max_int_key = -1
+        self._max_node = -1
         self.attributes = {'name': 'Graph'}
         self.node = {}
         self.edge = {}
@@ -142,69 +80,67 @@ class Graph(Datastructure):
     # --------------------------------------------------------------------------
 
     @property
+    def DATASCHEMA(self):
+        import schema
+        return schema.Schema({
+            "attributes": dict,
+            "dna": dict,
+            "dea": dict,
+            "node": dict,
+            "edge": dict,
+            "adjacency": dict,
+            "max_node": schema.And(int, lambda x: x >= -1)
+        })
+
+    @property
     def data(self):
         """Return a data dict of this data structure for serialization.
         """
-        version = LooseVersion(compas.__version__)
-        meta = {
-            "compas": version.vstring.split('-')[0],
-            "datatype": self.dtype,
-            "data": None
-        }
         data = {
             'attributes': self.attributes,
-            'node_attributes': self.default_node_attributes,
-            'edge_attributes': self.default_edge_attributes,
+            'dna': self.default_node_attributes,
+            'dea': self.default_edge_attributes,
             'node': {},
             'edge': {},
             'adjacency': {},
-            'max_int_key': self._max_int_key
+            'max_node': self._max_node
         }
         for key in self.node:
             data['node'][repr(key)] = self.node[key]
-
         for u in self.edge:
             ru = repr(u)
             data['edge'][ru] = {}
             for v in self.edge[u]:
                 rv = repr(v)
                 data['edge'][ru][rv] = self.edge[u][v]
-
         for u in self.adjacency:
             ru = repr(u)
             data['adjacency'][ru] = {}
             for v in self.adjacency[u]:
                 rv = repr(v)
                 data['adjacency'][ru][rv] = None
-
-        if version < LooseVersion('0.16.5'):
-            return data
-        meta['data'] = data
-        return meta
+        return data
 
     @data.setter
     def data(self, data):
-        if 'compas' in data:
-            version = LooseVersion(compas.__version__)
-            if version < LooseVersion('0.16.5'):
-                raise Exception('The data was generated with an incompatible newer version of COMPAS: {}'.format(version.vstring.split('-')[0]))
+        if 'data' in data:
             data = data['data']
-
         attributes = data.get('attributes') or {}
         default_node_attributes = data.get('dna') or {}
         default_edge_attributes = data.get('dea') or {}
         node = data.get('node') or {}
         edge = data.get('edge') or {}
         adjacency = data.get('adjacency') or {}
-
-        self._max_int_key = data.get('max_int_key')
+        if 'max_int_key' in data:
+            max_node = data['max_int_key']
+        else:
+            max_node = data.get('max_node')
+        self._max_node = max_node
         self.attributes.update(attributes)
         self.default_node_attributes.update(default_node_attributes)
         self.default_edge_attributes.update(default_edge_attributes)
-
         # add the nodes
         self.node = {literal_eval(key): attr for key, attr in iter(node.items())}
-
         # add the edges
         self.edge = {}
         for u, nbrs in iter(edge.items()):
@@ -215,7 +151,6 @@ class Graph(Datastructure):
                 attr = attr or {}
                 v = literal_eval(v)
                 self.edge[u][v] = attr
-
         # add the adjacency
         self.adjacency = {}
         for u, nbrs in iter(adjacency.items()):