+++ src/compas/geometry/primitives/polyline.py
@@ -64,7 +64,7 @@ class Polyline(Primitive):
         self.points = points
 
     @property
-    def dataschema(self):
+    def DATASCHEMA(self):
         from schema import Schema
         from compas.data import is_float3
         return Schema({