+++ src/compas/geometry/primitives/vector.py
@@ -81,7 +81,7 @@ class Vector(Primitive):
         self.z = z
 
     @property
-    def dataschema(self):
+    def DATASCHEMA(self):
         from schema import Schema
         from compas.data import is_float3
         return Schema(is_float3)